Add fix-presets script
Yamaha gives the first part of Montage preset performances the same name as
the whole performance, but subsequent parts use their instrument category in
upper-case as name. It is cleaner and more useful to rename these after the
performance too. No information is lost as each part already includes a
separate instrument category field.

fix-presets fixes all the active part names in presets to match the name of
the performance. This also corrects a handful of presets which were renamed
or corrected before release, but whose part names weren't updated in line.