Add first draft of file format top-level structure
To avoid this becoming too unwieldy a document, more detailed notes on
the contents of performance and live set data blocks will be broken out
into separate pages.

Known omissions in this early version include the library info format
and the meaning of the flag bytes in song and pattern entries.