Add a tool to rescale volumes of performances and live sets
This is a practical utility as well as an easy example, and is a rewrite
of my older shell script with the same function.

Many third-party performance libraries are normalised to noticeably
higher volumes than the presets, a habit which can be annoying (or
deafening) when auditioning a combined collection.

It is easy to fix a performance by editing its volume and resaving, but
this is time-consuming across a large library. Instead, run

  fixvolume SCALE-FACTOR [FILE]...

to scale all the performance and live set volumes in the given file(s)
by a constant factor specified in percentage ("85%"), decimal ("1.05")
or fractional ("80/67") form.