Add pruneblk to strip unneeded content out of files
pruneblk TYPE[,TYPE]... [FILE]... removes the specified block types.
TYPE is not case-sensitive and should be 'xxx' to remove 'Exxx' and
'Dxxx' blocks, or 'LIB' to clear installed library info.

Pruning system settings is useful to prevent them being clobbered on
reload of user or backup files, and standalone files without external
dependencies are tidier without the clutter of unused library info.

As no part of the file can grow and its blocks cannot overlap, we modify
and compact it in-place using memmove() on a writeable memory mapping.